Refactor: [사용자 관리] 루틴 리포트 생성 및 조회 보완 (MOYEO-77)
- RestTemplate 빈 주입 설정 추가
- 배치 확인 로직 루프 변경
- 하드코딩 제거
- ArrayList -> List 로 구체 컬렉션 의존 제거
- Gemini 응답 자체가 없을 경우 예외 처리 추가
