+++ src/main/java/com/moyeo/backend/routine/domain/RoutineReport.java
@@ -22,7 +22,7 @@ public class RoutineReport extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.UUID)
     private String id;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
