+++ src/main/java/com/moyeo/backend/routine/domain/RoutineReportRepository.java
@@ -1,11 +1,11 @@
 package com.moyeo.backend.routine.domain;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 public interface RoutineReportRepository {
-    void upsertAll(ArrayList<RoutineReport> reports);
+    void upsertAll(List<RoutineReport> reports);
 
     Optional<RoutineReport> findByUserIdAndStartDateAndIsDeletedFalse(String userId, LocalDate startDate);
 }
