--- src/main/java/com/moyeo/backend/routine/infrastructure/RoutineReportUpsertRepository.java
@@ -1,9 +1,0 @@
-package com.moyeo.backend.routine.infrastructure;
-
-import com.moyeo.backend.routine.domain.RoutineReport;
-
-import java.util.ArrayList;
-
-public interface RoutineReportUpsertRepository {
-    void upsertAll(ArrayList<RoutineReport> reports);
-}
