+++ src/main/java/com/moyeo/backend/routine/infrastructure/repository/RoutineReportUpsertRepository.java
@@ -1,0 +1,9 @@
+package com.moyeo.backend.routine.infrastructure.repository;
+
+import com.moyeo.backend.routine.domain.RoutineReport;
+
+import java.util.List;
+
+public interface RoutineReportUpsertRepository {
+    void upsertAll(List<RoutineReport> reports);
+}
