+++ src/main/java/com/moyeo/backend/routine/presentation/RoutineControllerDocs.java
@@ -7,6 +7,7 @@ import com.moyeo.backend.routine.application.dto.RoutineStatReadResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.http.ResponseEntity;
 
 @Tag(name = "학습 통계 관련 API Controller", description = "학습 통계 관련 API 목록입니다.")
@@ -17,6 +18,6 @@ public interface RoutineControllerDocs {
     ResponseEntity<ApiResponse<RoutineStatReadResponseDto>> getRoutineStat(RoutineReadRequestDto requestDto);
 
     @Operation(summary = "본인 루틴 리포트 조회 API", description = "본인 루틴 리포트 조회 API 입니다.")
-    ResponseEntity<ApiResponse<RoutineReportReadResponseDto>> getRoutineReport(RoutineReadRequestDto requestDto);
+    ResponseEntity<ApiResponse<RoutineReportReadResponseDto>> getRoutineReport(@ParameterObject RoutineReadRequestDto requestDto);
 
 }
