+++ src/main/resources/application.yml
@@ -77,6 +77,7 @@ oauth:
 
 ai:
   gemini:
+    name: GEMINI
     base-url: https://generativelanguage.googleapis.com
     model: gemini-2.5-flash
     key: ${GEMINI_API_KEY}
